+++ CHANGES.md
@@ -3,6 +3,13 @@
 
 * Latest Hackage released version: 8.6, 2020-02-08
 
+### Version 8.7, Not yet released
+
+  * Add support for concurrent versions of solvers for query
+    problems. Similar to 'satWithAny', 'proveWithAny' etc.,
+    except when we have queries. Thanks to Jeffrey Young
+    for the idea and the implementation.
+
 ### Version 8.6, 2020-02-08
 
   * Fix typo in error message. Thanks to Oliver Charles
@@ -394,7 +401,7 @@
         * Existential  : bAny    became   sAny
         * Universal    : bAll    became   sAll
 
-  * [BACKWARDS COMPATIBILITY, INTERNAL] Hostorically, SBV focused on bit-vectors and machine
+  * [BACKWARDS COMPATIBILITY, INTERNAL] Historically, SBV focused on bit-vectors and machine
     words, which meant lots of internal types were named suggestive of this heritage.
     With the addition of `SInteger`, `SReal`, `SFloat`, `SDouble` we have expanded
     this, but still remained focused on atomic types. But, thanks largely to